raw-replay/requirements.txt
quixstreams<2.9
python-dotenv
pandas
Faker
structlog
